requirements.txt
groq
yfinance
pandas
numpy
pydantic
python-dotenv
pytest
